\name{hyperSMURF-package}
\alias{hyperSMURF-package}
\alias{hyperSMURF}
\docType{package}
\title{
\packageTitle{hyperSMURF}
}
\description{
\packageDescription{hyperSMURF}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{hyperSMURF}
\packageIndices{hyperSMURF}
}
\author{
\packageAuthor{hyperSMURF}

Maintainer: \packageMaintainer{hyperSMURF}
}

\references{
M. Schubach, M. Re, P.N. Robinson and G. Valentini Imbalance-Aware Machine Learning for Predicting Rare and Common Disease-Associated Non-Coding Variants,
Scientific Reports, Nature Publishing, 7:2959, 2017.
}
