\name{hyperSMURF-package}
\alias{hyperSMURF-package}
\alias{hyperSMURF}
\docType{package}
\title{
\packageTitle{hyperSMURF}
}
\description{
\packageDescription{hyperSMURF}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{hyperSMURF}
\packageIndices{hyperSMURF}
}
\author{
\packageAuthor{hyperSMURF}

Maintainer: \packageMaintainer{hyperSMURF}
}
